Move `compare_metrics` and `plot_overall_rank_comparison` to `Evaluator`
