+++ tabrepo/__init__.py
@@ -1,3 +1,4 @@
 from .repository.evaluation_repository import EvaluationRepository, load_repository
 from .repository.evaluation_repository_collection import EvaluationRepositoryCollection
 from .contexts._context_fetcher import get_context, list_contexts
+from .evaluation.evaluator import Evaluator
